% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{arrangeTopGeneraInDonor}
\alias{arrangeTopGeneraInDonor}
\title{Top genera ranked based on the Donor group
#'
Arrange the top genera based on the Donor group}
\usage{
arrangeTopGeneraInDonor(
  projectmetadata_w_tax_hellinger_transformed,
  top_genera
)
}
\arguments{
\item{projectmetadata_w_tax_hellinger_transformed}{A df created by \code{hellingerTransformSamples}}

\item{top_genera}{A df created by \code{calculateTopGenera}}
}
\description{
Top genera ranked based on the Donor group
#'
Arrange the top genera based on the Donor group
}
