% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\name{calculateSpeciesRichness}
\alias{calculateSpeciesRichness}
\title{Calculate the species richness
#'
Calculate the species richness defined as number of species present in a sample. Filter_species will first filter all species with less than x relative abundance in at least one sample}
\usage{
calculateSpeciesRichness(metaphlan, filter_species = 0)
}
\arguments{
\item{metaphlan}{A metaphlan dataframe}

\item{filter_species}{A cutoff value in relative abundance.}
}
\description{
Calculate the species richness
#'
Calculate the species richness defined as number of species present in a sample. Filter_species will first filter all species with less than x relative abundance in at least one sample
}
