% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{calculateTopGenera}
\alias{calculateTopGenera}
\title{Finding top genera
#'
finds the top genera based on weighted mean of hellinger transformed relative abundances.}
\usage{
calculateTopGenera(
  projectmetadata_w_tax_hellinger_transformed,
  stages = c("Pre", "Donor"),
  n_genera
)
}
\arguments{
\item{projectmetadata_w_tax_hellinger_transformed}{A df created by \code{hellingerTransformSamples}}

\item{stages}{A vector with stages which weighted mean will be based on. Defaults to \code{c("Pre", "Donor")}}

\item{n_genera}{Top n genera.}
}
\description{
Finding top genera
#'
finds the top genera based on weighted mean of hellinger transformed relative abundances.
}
