% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{clusterPreAndDonorSamples}
\alias{clusterPreAndDonorSamples}
\title{Cluster samples for heatmap
#'
calculates clusters based on hellinger transformed bray-curtis distance with the ward.D2 algorithm.}
\usage{
clusterPreAndDonorSamples(metadata_tax_hellinger, t_metaphlan)
}
\arguments{
\item{metadata_tax_hellinger}{A df made by \code{hellingerTransformSamples}}

\item{t_metaphlan}{A metaphlan df transposed}
}
\description{
Cluster samples for heatmap
#'
calculates clusters based on hellinger transformed bray-curtis distance with the ward.D2 algorithm.
}
