% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{compareFMT2ActualDonor}
\alias{compareFMT2ActualDonor}
\title{Get FMT-DONOR comparisons}
\usage{
compareFMT2ActualDonor(projectmetadata, metadata_beta)
}
\arguments{
\item{projectmetadata}{A df with project related metadata created by \code{isolateProjectMetadata}}

\item{metadata_beta}{A df created by \code{createMetadataWBetadiversity}}
}
\description{
extracts the comparison between the FMT group and actual donors received.
}
