% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{compareFMT2DonorNotReceived}
\alias{compareFMT2DonorNotReceived}
\title{Get FMT-DONOR_not_received comparisons
#'
extracts the comparison between the FMT group and donors not received.}
\usage{
compareFMT2DonorNotReceived(projectmetadata, metadata_beta)
}
\arguments{
\item{projectmetadata}{A df with project related metadata created by \code{isolateProjectMetadata}}

\item{metadata_beta}{A df created by \code{createMetadataWBetadiversity}}
}
\description{
Get FMT-DONOR_not_received comparisons
#'
extracts the comparison between the FMT group and donors not received.
}
