% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{comparePlacebo2DonorNotReceived}
\alias{comparePlacebo2DonorNotReceived}
\title{Compare placebo group to donor group}
\usage{
comparePlacebo2DonorNotReceived(projectmetadata, metadata_beta)
}
\arguments{
\item{projectmetadata}{A df with project related metadata created by \code{isolateProjectMetadata}}

\item{metadata_beta}{A df created by \code{createMetadataWBetadiversity}}
}
\description{
extracts the comparison between the placebo group and donors.
}
