% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{createANDcombineComparedOutputs}
\alias{createANDcombineComparedOutputs}
\title{Bind comparisons}
\usage{
createANDcombineComparedOutputs(projectmetadata, metadata_beta)
}
\arguments{
\item{projectmetadata}{A df from \code{isolateProjectMetadata}}

\item{metadata_beta}{A df from \code{createMetadataWBetadiversity}}
}
\description{
A helper function to bind_rows of the compared dataframes
}
