% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{createBetaDistMatrix}
\alias{createBetaDistMatrix}
\title{createBetaDistMatrix}
\usage{
createBetaDistMatrix(t_metaphlan, method = "bray", transform = T)
}
\arguments{
\item{t_metaphlan}{A metaphlan df transposed}

\item{method}{Either bray or sorensen}

\item{transform}{Boolean. Should hellinger transformation be applied or not?}
}
\description{
creates beta dissimilarity matrix
}
