% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{createMetadataWBetadiversity}
\alias{createMetadataWBetadiversity}
\title{Betadiversity dissimilarity with metadata}
\usage{
createMetadataWBetadiversity(
  projectmetadata,
  t_metaphlan,
  method = "bray",
  transform = TRUE
)
}
\arguments{
\item{projectmetadata}{A metadata df that has been processed by \code{isolateProjectAndDonor}}

\item{t_metaphlan}{A transposed metahplan df.}

\item{method}{specifies which beta-distance should be measured. Choose between "bray" or "sorensen". Defaults to bray.}

\item{transform}{Boolean. Specifies if metaphlan df should be hellinger-transformed before distance calculation}
}
\description{
Function creates a distance matrix, transform the matrix to a dataframe and adds projectmetadata to the dataframe.
}
