% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\name{filter_taxonomy}
\alias{filter_taxonomy}
\title{Metaphlan taxonomy filter}
\usage{
filter_taxonomy(
  metaphlan,
  taxonomy = "species",
  keep_unclassified = TRUE,
  remove_tax_names = TRUE
)
}
\arguments{
\item{metaphlan}{}

\item{taxonomy}{}

\item{keep_unclassified}{}
}
\description{
Metaphlan taxonomy filter
}
