% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg_alpha}
\alias{gg_alpha}
\title{Title
#'
Description}
\usage{
gg_alpha(df, x, alpha_metric = "richness")
}
\arguments{
\item{df}{A dataframe containing the columns x, alpha-diversity metric}

\item{x}{A vector with the variables from the x_axis column which will be plotted on the x-axis. The vector must be given in the order to appear on the x-axis.}

\item{alpha_metric}{A variable stating the name of the column, which contains the alpha-diversity metric.}
}
\description{
Title
#'
Description
}
