% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{gg_beta}
\alias{gg_beta}
\title{Similiarity to donor plot
#'
Creates a similarity to donor plot.}
\usage{
gg_beta(metadata_beta_summarised, method)
}
\arguments{
\item{metadata_beta_summarised}{A summarised beta diversity df}

\item{method}{Which method was used (either sorensen or bray)}
}
\description{
Similiarity to donor plot
#'
Creates a similarity to donor plot.
}
