% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{hellingerTransformSamples}
\alias{hellingerTransformSamples}
\title{Hellinger transform metadata with taxonomic data}
\usage{
hellingerTransformSamples(projectmetadata_w_tax)
}
\arguments{
\item{projectmetadata_w_tax}{A df with metadata and taxonomic data.}
}
\description{
Helper function in generating the y-axis for the heatmap. Each sample batch is hellinger transformed
}
