% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{isolateProjectMetadata}
\alias{isolateProjectMetadata}
\title{Isolates metadata for project to be analyzed}
\usage{
isolateProjectMetadata(metadata, project_filter = "MP")
}
\arguments{
\item{metadata}{A dataframe}

\item{project_filter}{A project}
}
\description{
Functionen for isolating relevant metadata. Three projects can be used: MP or NP
}
