% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\name{load_metaphlan}
\alias{load_metaphlan}
\title{Loads a metaphlan table}
\usage{
load_metaphlan(taxonomic_level = "genus")
}
\arguments{
\item{taxonomic_level}{A taxonomic level, which loads the corresponding metaphlan table}
}
\description{
Loads a metaphlan table filtered to a certain taxonomic level
}
