% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\name{remove_NonHuman_from_colnames}
\alias{remove_NonHuman_from_colnames}
\title{Rename columns in metaphlan}
\usage{
remove_NonHuman_from_colnames(metaphlan)
}
\arguments{
\item{metaphlan}{}
}
\description{
Rename columns in metaphlan
}
