% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{summariseBetaDiversityOutput}
\alias{summariseBetaDiversityOutput}
\title{Summarise betadiversity
#'
calculates the median dissimiliarity of each sample to donors received and not received.}
\usage{
summariseBetaDiversityOutput(metadata_beta_compared)
}
\arguments{
\item{metadata_beta_compared}{A df created by one of the \code{compare...} functions}
}
\description{
Summarise betadiversity
#'
calculates the median dissimiliarity of each sample to donors received and not received.
}
