% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan.R
\name{transposeMetaphlan}
\alias{transposeMetaphlan}
\title{transpose metaphlan table
#'
transpose metaphlan table}
\usage{
transposeMetaphlan(metaphlan)
}
\arguments{
\item{metaphlan}{A metaphlan dataframe}
}
\description{
transpose metaphlan table
#'
transpose metaphlan table
}
