% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadiversity.R
\name{createMetadataBetaWDonorReceived}
\alias{createMetadataBetaWDonorReceived}
\title{Donor received}
\usage{
createMetadataBetaWDonorReceived(projectmetadata, metadata_beta)
}
\arguments{
\item{projectmetadata}{A df with project related metadata created by \code{isolateProjectMetadata}}

\item{metadata_beta}{A df created by \code{createMetadataWBetadiversity}}
}
\description{
Helper function: Adds the donorbatches the patient received during treatment to metadata_beta
}
